JSON Library and Radial Flipping
ADDED:
- Json Library re-implemented, will be used for loading spawner configs for level
- RadialSpawner can now flip direction every iteration
- RadialSpawner can now properly start reversed

NOTE:
Still seems to be a bug with split spreams creating a line at 90 degrees

+++ Warptex/RadialSpawner.h
@@ -26,7 +26,7 @@ protected:
 	std::vector<RadialPattern> patternQueue = std::vector<RadialPattern>();
 
 	float amountFired = 0;
-	float iterationsPassed = 0;
+	int iterationsPassed = 0;
 	bool reversing = false;
 	bool active = false;
 };
